changes as suggested by viboe: • renamed executor to differentate different versions
